Major Update 0.3: Built-In 'Your Cards' File Explorer.
Flash-Carder:
- Switched the finish command from 'Done' to 'Enter'
- Now mentions if 2 or more answers/parts to a question.
- Removed super inaccurate marking opinions.
- Added 'Another Card Data File.txt'

New 'Your Cards' File Explorer:
- Option to either...
Load the 'default' (hard-coded in) card data file, or...
Pick another from the 'Your Cards' folder.

...All within the Carder program.
(Only available if 2 or more card data files are held in the 'Your Cards' folder.)

- Pick your file using the appropriate file number from the file 'tree' display.
- No need to hard-code every time now. Have one default, and pick others when necessary.

JC:
- Increased 'dramatypeSpeed' and general pacing speed.

- Added 'Digits' lambda expression:
Returns amount of digits of a number.

- Added 'Gapped' arguement to 'IsBusy' function:
Toggles newlines on/off.

- Added 'RangeCheck' function:
Verifies if a value is in range, and raises a 'FoundError' if not.